requirements.txt
telethon
requests
aiohttp
pillow
bs4
webcolors
gtts
art
pyfiglet
whois
rarfile
hachoir
qrcode
yt-dlp
wikipedia
langdetect
beautifulsoup4
phonenumbers
piexif
unidecode
pyzbar
faker
googletrans==4.0.0-rc1